+++ src/Render.h
@@ -3,7 +3,7 @@
 
 #include "GameTypes.h"
 
-const static size_t RENDER_QUEUE_SIZE = 4096;
+const static int RENDER_QUEUE_SIZE = 4096;
 
 namespace Render
 {
@@ -26,10 +26,10 @@ struct RenderRectangleEvent
 
 struct RenderTextureEvent
 {
-    size_t texture_index;
+    int texture_index;
     Rect clip;
     V2 position;
-    size_t scale;
+    int scale;
     bool has_clip;
 };
 struct Event
@@ -38,15 +38,15 @@ struct Event
     Render::EventType type;
     Rect overflow_clip;
     bool has_overflow_clip;
-    size_t z_index;
+    int z_index;
     union {
         Render::RenderRectangleEvent render_rectangle_event;
         Render::RenderTextureEvent render_texture_event;
     } data;
 };
-void render_texture(Render::Layer layer, size_t texture_index, V2 &position, Rect *overflow_clip = nullptr, size_t scale = 1, size_t z_index = 1);
-void render_texture(Render::Layer layer, size_t texture_index, Rect &clip, V2 &position, Rect *overflow_clip = nullptr, size_t scale = 1, size_t z_index = 1);
-void render_rectangle(Render::Layer layer, const Rect &box, const Color &color, bool filled = false, size_t z_index = 1);
+void render_texture(Render::Layer layer, int texture_index, V2 &position, Rect *overflow_clip = nullptr, int scale = 1, int z_index = 1);
+void render_texture(Render::Layer layer, int texture_index, Rect &clip, V2 &position, Rect *overflow_clip = nullptr, int scale = 1, int z_index = 1);
+void render_rectangle(Render::Layer layer, const Rect &box, const Color &color, bool filled = false, int z_index = 1);
 void perform_render();
 }; // namespace Render
